feat: #798 - simplification through abstraction
New files:
* `abstract_user_preferences.dart`: Abstraction of a collapsed/expanded display for the preferences page.
* `user_preferences_attribute_group.dart`: Collapsed/expanded display of attributes for the preferences page.
* `user_preferences_profile.dart`: Collapsed/expanded display of profile for the preferences page.
* `user_preferences_settings.dart`: Collapsed/expanded display of settings for the preferences page.

Impacted file:
* `user_preferences_page.dart`: huge simplification through the use of instances of new class `AbstractUserPreferences`
