+++ packages/smooth_app/lib/pages/user_preferences_settings.dart
@@ -1,0 +1,200 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import 'package:smooth_app/helpers/launch_url_helper.dart';
+import 'package:smooth_app/pages/abstract_user_preferences.dart';
+import 'package:smooth_app/themes/smooth_theme.dart';
+import 'package:smooth_app/themes/theme_provider.dart';
+import 'package:smooth_app/views/bottom_sheet_views/user_contribution_view.dart';
+import 'package:smooth_ui_library/buttons/smooth_simple_button.dart';
+import 'package:smooth_ui_library/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_ui_library/widgets/smooth_list_tile.dart';
+import 'package:smooth_ui_library/widgets/smooth_toggle.dart';
+
+/// Collapsed/expanded display of settings for the preferences page.
+class UserPreferencesSettings extends AbstractUserPreferences {
+  UserPreferencesSettings(final Function(Function()) setState)
+      : super(setState);
+
+  static const List<String> _ORDERED_COLOR_TAGS = <String>[
+    SmoothTheme.COLOR_TAG_BLUE,
+    SmoothTheme.COLOR_TAG_GREEN,
+    SmoothTheme.COLOR_TAG_BROWN,
+  ];
+
+  @override
+  bool isCollapsedByDefault() => true;
+
+  @override
+  String getPreferenceFlagKey() => 'settings';
+
+  @override
+  String getTitle() => 'App Settings';
+
+  @override
+  String getSubtitle() => 'Dark mode, country, color, ...';
+
+  @override
+  List<Widget> getBody(
+    final BuildContext context,
+    final AppLocalizations appLocalizations,
+    final ThemeProvider themeProvider,
+    final ThemeData themeData,
+  ) =>
+      <Widget>[
+        SmoothListTile(
+          text: appLocalizations.darkmode,
+          onPressed: null,
+          leadingWidget: SmoothToggle(
+            value: themeProvider.darkTheme,
+            width: 85.0,
+            height: 38.0,
+            textRight: appLocalizations.darkmode_light,
+            textLeft: appLocalizations.darkmode_dark,
+            colorRight: Colors.blue,
+            colorLeft: Colors.blueGrey.shade700,
+            iconRight: const Icon(Icons.wb_sunny_rounded),
+            iconLeft: const Icon(
+              Icons.nightlight_round,
+              color: Colors.black,
+            ),
+            onChanged: (bool newValue) async =>
+                themeProvider.setDarkTheme(newValue),
+          ),
+        ),
+        SmoothListTile(
+          leadingWidget: Container(),
+          title: Wrap(
+            spacing: 8.0,
+            children: List<Widget>.generate(
+              _ORDERED_COLOR_TAGS.length,
+              (final int index) => _getColorButton(
+                themeData.colorScheme,
+                _ORDERED_COLOR_TAGS[index],
+                themeProvider,
+              ),
+            ),
+          ),
+        ),
+        SmoothListTile(
+          text: appLocalizations.contribute,
+          onPressed: () => showCupertinoModalBottomSheet<Widget>(
+            expand: false,
+            context: context,
+            backgroundColor: Colors.transparent,
+            bounce: true,
+            builder: (BuildContext context) => UserContributionView(),
+          ),
+        ),
+        SmoothListTile(
+          text: appLocalizations.support,
+          leadingWidget: const Icon(Icons.launch),
+          onPressed: () => LaunchUrlHelper.launchURL(
+              'https://slack.openfoodfacts.org/', false),
+        ),
+        SmoothListTile(
+          text: appLocalizations.about_this_app,
+          onPressed: () async {
+            final PackageInfo packageInfo = await PackageInfo.fromPlatform();
+            showDialog<void>(
+              context: context,
+              builder: (BuildContext context) => SmoothAlertDialog(
+                close: false,
+                body: Column(
+                  children: <Widget>[
+                    ListTile(
+                      leading:
+                          Image.asset('assets/app/smoothie-icon.1200x1200.png'),
+                      title: Text(
+                        packageInfo.appName,
+                        style: themeData.textTheme.headline1,
+                      ),
+                      subtitle: Text(
+                        packageInfo.version,
+                        style: themeData.textTheme.subtitle2,
+                      ),
+                    ),
+                    Divider(color: themeData.colorScheme.onSurface),
+                    const SizedBox(height: 20),
+                    Text(appLocalizations.whatIsOff),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        TextButton(
+                          onPressed: () => LaunchUrlHelper.launchURL(
+                              'https://openfoodfacts.org/who-we-are', true),
+                          child: Text(
+                            appLocalizations.learnMore,
+                            style: const TextStyle(
+                              color: Colors.blue,
+                            ),
+                          ),
+                        ),
+                        TextButton(
+                          onPressed: () => LaunchUrlHelper.launchURL(
+                              'https://openfoodfacts.org/terms-of-use', true),
+                          child: Text(
+                            appLocalizations.termsOfUse,
+                            style: const TextStyle(
+                              color: Colors.blue,
+                            ),
+                          ),
+                        )
+                      ],
+                    )
+                  ],
+                ),
+                actions: <SmoothSimpleButton>[
+                  SmoothSimpleButton(
+                    onPressed: () async {
+                      showLicensePage(
+                        context: context,
+                        applicationName: packageInfo.appName,
+                        applicationVersion: packageInfo.version,
+                        applicationIcon: Image.asset(
+                          'assets/app/smoothie-icon.1200x1200.png',
+                          height: MediaQuery.of(context).size.height * 0.1,
+                        ),
+                      );
+                    },
+                    text: appLocalizations.licenses,
+                    minWidth: 100,
+                  ),
+                  SmoothSimpleButton(
+                    onPressed: () =>
+                        Navigator.of(context, rootNavigator: true).pop(),
+                    text: appLocalizations.okay,
+                    minWidth: 100,
+                  ),
+                ],
+              ),
+            );
+          },
+        ),
+      ];
+
+  Widget _getColorButton(
+    final ColorScheme colorScheme,
+    final String colorTag,
+    final ThemeProvider themeProvider,
+  ) =>
+      TextButton(
+        onPressed: () async => themeProvider.setColorTag(colorTag),
+        style: TextButton.styleFrom(
+          backgroundColor: SmoothTheme.getColor(
+            colorScheme,
+            SmoothTheme.MATERIAL_COLORS[colorTag]!,
+            ColorDestination.BUTTON_BACKGROUND,
+          ),
+        ),
+        child: Icon(
+          Icons.palette,
+          color: SmoothTheme.getColor(
+            colorScheme,
+            SmoothTheme.MATERIAL_COLORS[colorTag]!,
+            ColorDestination.BUTTON_FOREGROUND,
+          ),
+        ),
+      );
+}
